% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_locs.R
\name{combine.locs}
\alias{combine.locs}
\title{Re-merge quantile map projections}
\usage{
combine.locs(defaults, comments = character())
}
\arguments{
\item{defaults}{the output from \code{\link{various_defaults}}}

\item{comments}{if desired, a 'comments' global attribute will be
added to the NetCDF file, with the string input to this
option as the body}
}
\description{
Concatenate the individual projection files calculated through separate calls
to \code{build.projections} into a single file, removing duplicates along the
way.
}
\section{Output Files}{

The file is saved as both an .RData file and a NetCDF file. The final output
in the NetCDF file is a \code{[loc x time]} array (variable name
\code{[defaults$filevar]}), with variables \code{lat} and \code{lon} being
\code{[loc x 1]} vectors giving the location. The file also includes a variable
\code{base_idx} giving, for each projected day, the linear index in the
original base data that was used as the base day for the projection.
}

\section{Beware}{

Files are searched for in the directory code{[base.data.dir]/output}, with
\code{[base.data.dir]} set by \code{defaults}. As a result, this code depends
on the desired temporary files being the only files in that directory with
the filename pattern searched for (from the \code{file.list} variable in the
code).
}

