% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_params.R
\name{combine.params}
\alias{combine.params}
\title{Re-merge Quantile Regression Coefficients}
\usage{
combine.params(defaults)
}
\arguments{
\item{defaults}{the output from \code{\link{various_defaults}}}
}
\value{
a merged file containing all calculated regression coefficients in a
  large list, \code{params}, and saved in \code{[defaults$mod.data.dir]/params/}
  as \code{[filevar]_day_[mod.name]_quantfit_params_[mod.year.range[1]-mod.year.range[2]]_alllocs.RData}.
}
\description{
Concatenate the individual parameter files calculated through separate calls
to \code{\link{estimate.quantiles}} into a single file, removing duplicates along the
way. This is only really necessary for \code{\link{get.quantile.changes}}.
}
