% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncdf.R
\name{get.ncdf}
\alias{get.ncdf}
\title{Load Subsets of NetCDF Files By Location, Run, Time}
\usage{
get.ncdf(defaults, process.inputs.tmp, year.range = numeric())
}
\arguments{
\item{defaults}{the output from \code{\link{set.defaults}}. The defaults 
used are \code{filevar} and \code{varnames} to search for grid/run  
variablesin the files.}

\item{process.inputs.tmp}{the output from \code{\link{get.process.chunks}},
giving the information needed to load a chunk of data from NetCDF file(s), 
including filename(s), locations within files, etc.}

\item{year.range}{which year range should be output, in a vector 
\code{c(year_i,year_f)}. If nothing set, the full possible year range of 
the file is loaded.}
}
\value{
A list; each element reperesenting a pixel, containing named
 subelements \code{lat}, \code{lon}, and the raw data \code{Raw}.
}
\description{
\code{get.ncdf} loads the data stored in NetCDF files, as determined by 
 the information in \code{process.inputs.tmp}, output from 
 \code{\link{get.process.chunks}}. Data is loaded by region, latitude 
 band, run, and time range, and is concatenated across files if necessary.
 Data is returned as a list by pixel, with every list member containing 
 the \code{lat} and \code{lon} of the pixel, and the raw data 
 ("\code{Raw}").
}
\section{Expected File Structure}{

 In general, most common forms of climate file structures are supported, 
 especially the CMIP5 structure. Variables can either be on a \code{lon x lat} 
 grid or stored by linear location. Files can either contain all runs of a model
 or can be saved by run. Files can either contain the whole timeframe of a model
 run or be split up in consecutive temporal chunks. Furthermore:
 \describe{
   \item{filename}{the code searches for \emph{NetCDF} files using the search
     string "\code{[defaults$filevar]_day_.*nc}" (by default; this can be 
   changed by setting \code{defaults$search.str}). Make sure no other NetCDF
     files with that pattern are present in the search directory (by default
     \code{defaults$mod.data.dir}).}
   \item{variable setup}{Currently, the code expects the primary variable to 
   have either a location dimension (giving the linear index of a location),
   or a lon x lat grid. These are all identified by name - the search terms
   used can be set in \code{defaults$varnames} - out-of-the-box, the package 
   for example supports "lat", "latitude", "Latitude", and "latitude_1" as 
   possible names for the "lat" dimension.}
   \item{locations}{The code expects there to be two location variables,
     \code{lat} and \code{lon} (CMIP5 syntax), giving the lat/lon location of
      every pixel in the file. The names of those variables can be any of the
    alternatives given by \code{defaults$varnames}.}
   \item{multiple runs}{If there are multiple runs in the file, there should
     be a \code{run} variable/dimension in the file giving the run id as an
     integer}
 }
}

\section{Warnings and Notes on File Structure}{
 
 \itemize{
   \item Files have to begin on 01/01 and end on 12/31 of a year. 
   \item Loading speeds up a lot if as much of the data as possible is stored
     in a single file - an example is storing files by region, but with all
     runs and years in the same file for the pixels in that region.
 }
}

