% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictors.R
\name{get.predictors}
\alias{get.predictors}
\title{Build spline basis functions}
\usage{
get.predictors(n_files = 1, dfs = numeric(), df.x = numeric(),
  df.t = numeric(), df.xt = numeric(), year.range = c(1850, 2099),
  get.volc = FALSE, lat = numeric(), save.predictors = FALSE,
  save.fn = character())
}
\arguments{
\item{n_files}{number of runs (if \code{> 1},
this ensures that Jan 1, 1971, Run
1 has the same basis function value
as Jan 1, 1971, Run 2, etc.). Basis
functions are repeated end-to-end
\code{n_files} times.}

\item{dfs}{degrees of freedom - a \code{[3 x 1]} vector giving the
seasonal, long term change, and interaction degrees of
freedom, respectively}

\item{df.x, df.t, df.xt}{alternatively, input the dfs separately,
with seasonal \code{df.x}, long-term \code{df.t}, and
interaction \code{df.xt} (kept for backwards compatibility
with some older functions)}

\item{year.range}{\code{c(start_year,end_year)}, or, if
\code{get.volc=FALSE}, can also just be a number giving
the length in years of the used time series.}

\item{get.volc}{whether to add a volcanic degree of freedom}

\item{lat}{only necessary if \code{get.volc=TRUE}; This is the
latitude of the pixel for which basis functions
should be calculated.}

\item{save.predictors}{if true, the basis functions are saved
to an .RData file (def: \code{FALSE})}

\item{save.fn}{if \code{save.predictors=TRUE}, this sets the name of the
file to save the basis functions to. By default, the
filename is: "\code{[aux.dir]/bases/spline_basis_functions_[nyears]years_[n_files]runs_[df.x]-[df.t]-[df.xt]df(_[lat.volc]volc).RData}"}
}
\value{
a numerical matrix
}
\description{
\code{get.predictors} builds basis functions as predictors for
the quantile regression in the quantile mapping climate
projection method of \emph{Haugen et al. 2018}. The basis
functions are constructed such that the resultant quantile
estimate is smoothly varying and depends on:
\enumerate{
   \item the seasonal cycle (dependent only on month-of-year),
         estimated by a periodic b-spline
   \item the long-term trend (dependent only on the year index),
         estimated by a natural cubic spline
   \item an interaction between the seasonal cycle and long-term
         trend (to capture changes in the seasonal cycle)
}
The degrees of freedom for each of these basis functions are set through the
parameters \code{df.x}, \code{df.t}, and \code{df.xt}, respectively, or through a \code{3 x 1} vector \code{dfs}.
}
\section{Multiple runs/files}{

\code{get.predictors} is specifically designed to allow for basis
functions for quantile regression across multiple data sets that
span the same time period. In the context of the quantile mapping
climate projection method, these multiple data sets are multiple
runs of the same model over the same time period. In this case, the
basis vectors are repeated end-to-end \code{n_files} times.
}

\section{Volcanic forcing}{

The \code{get.volc} input additionally allows for the addition
of a historical volcanic CO2 forcing as a predictor for the
quantile regression. This volcanic forcing is taken from the
XXXXXX hindcast from CCSM4XXXX. If \code{get.volc=TRUE}, then a
latitude must be specified through the \code{lat} parameter.
Since the volcanic forcing is only available 1850-2008, this
is discouraged for values beyond 2008...
}

