% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_process_chunks.R
\name{get.process.chunks}
\alias{get.process.chunks}
\title{Get File Information to Allow Processing of Subsets}
\usage{
get.process.chunks(defaults, save.output = FALSE, search.dir = character())
}
\arguments{
\item{defaults}{the output from \code{\link{set.defaults}}. The defaults used
are \code{filevar} and, if \code{search.dir=numeric()} (by default),
\code{mod.data.dir}.}

\item{save.output}{whether to save the file information as a file in the
search directory (as "\code{process_inputs.RData}"), by default
\code{FALSE}}

\item{search.dir}{by default, \code{get.process.chunks} searches in the
\code{mod.data.dir} in \code{defaults}. If a different directory should be
examined, use this to set the path.}
}
\value{
a list giving, for each region-by-latitude chunk the subset suffix
  (\code{reg}) if available, the filename (\code{fn}), the latitude and
  longitude coordinates of the pixels in the subset (\code{lat}, one element;
  and \code{lon}), the global pixel id (the variable \code{global_loc} if it
  exists in the \emph{NetCDF} file, otherwise a new one is created, counting
  pixels up by files alphabetically), the local id (\code{local_idxs}) linear
  index within the file of each pixel), and the number of experiment runs in
  the file (the variable \code{run} in the \emph{NetCDF} file; 1 otherwise).
}
\description{
Get information about the climate data files in a folder,
and split them up by region-by-latitude chunk to allow for
processing subsections of the (very large) files at a time.
}
\section{Expected file structure}{

 \code{get.process.chunks} and the \code{quantproj} packages as a whole
 expect the climate data to follow a few conventions:
 \describe{
   \item{filename}{the code searches for \emph{NetCDF} files using the search
     string "\code{[defaults$filevar]_day_.*nc}". Make sure no other NetCDF
     files with that pattern are present in the search directory (by default
     \code{defaults$mod.data.dir}).}
   \item{variable setup}{Currently, the code expects the primary variable to
     be a \code{[loc x time (x run)]} \emph{2-dimensional} variable named
     \code{defaults$filevar}. Future versions will support \code{[lon x lat x
     time (x run)]} grids as well. }
   \item{locations}{The code expects there to be two location variables,
     \code{lat} and \code{lon} (CMIP5 syntax), giving the lat/lon location of
      every pixel in the file}
   \item{multiple runs}{If there are multiple runs in the file, there should
     be a \code{run} variable/dimension in the file giving the run id as an
     integer}
 }
}

