% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various_defaults_func.R
\name{set.defaults}
\alias{set.defaults}
\title{Set default inputs for quantile mapping}
\usage{
set.defaults(aux.dir = "aux/", base.data.dir = "base_data/",
  mod.data.dir = "model_data/", filevar = "tas", mod.name = "LENS",
  base.name = "ERA-INTERIM", base.year.range = c(1979, 2010),
  mod.year.range = c(1979, 2099), proj.year.range = c(2011, 2099),
  q_norm = c(0.1, 0.5, 0.9), norm.x.df = c(14, 6, 3),
  base.norm.x.df = c(10, 1, 0), get.volc = F, q_bulk = c(0.1, 0.18, 0.25,
  0.35, 0.42, 0.5, 0.58, 0.65, 0.75, 0.82, 0.9), bulk.x.df = c(14, 6, 3),
  q_tail = c(0.01, 0.1, 0.25, 0.5, 0.75), tail.x.df = c(3, 1, 0),
  index.type = "resampling_rep", resampling.timescale = "year",
  bootstrapping = F, nboots = 50, block.size = 40)
}
\arguments{
\item{aux.dir}{directory for auxiliary files - saved basis functions, etc.}

\item{base.data.dir}{the directory with raw data to be projected}

\item{mod.data.dir}{the directory with raw model ensemble data used to project}

\item{filevar}{variable shorthand (CMIP5 syntax, def: \code{"tas"} for near-surface air temperature)}

\item{mod.name}{name of model ensemble (def: \code{"LENS"})}

\item{base.name}{name of data product to be projected (def: \code{"ERA-INTERIM"})}

\item{base.year.range}{desired time frame of the "base" time period to project (def: \code{c(1979,2010)})}

\item{mod.year.range}{desired time frame of the time frame to be processed (quantiles calculated) for the model ensmeble; make sure this range includes \code{base.year.range} (def: \code{c(1979,2099)})}

\item{proj.year.range}{desired time frame to project to (def: \code{c(2011,2099)})}

\item{q_norm}{normalizing quantiles (must be a \code{3 x 1} vector, i.e. \code{c(0.1,0.5,0.9)})}

\item{norm.x.df, bulk.x.df, tail.x.df}{degrees of freedom for normalizing, bulk, and tail quantile regressions, in a \code{3 x 1} vector for \code{c([seasonal],[long-term],[interaction])}}

\item{get.volc}{include a volcanic predictor in the normalization (def: \code{FALSE})}

\item{q_bulk}{bulk quantiles (must be a \code{[n_bulk_quantiles x 1]} vector, i.e. \code{c(0.25,0.3,0.5,0.6,0.75)})}

\item{q_tail}{tail quantiles (must be a \code{[n_bulk_quantiles x 1]} vector, i.e. \code{c(0.5,0.9,0.95)}). These are fractions of the difference between \code{max(q_bulk)} and \code{0}/\code{1}!}

\item{nboots}{number of bootstrap runs (def: \code{50})}

\item{block.size}{bootstrap block size (def: \code{40}) (ACROSS RUNS - SO 20 means each bootstrap block is across 20 randomly chosen runs)}

\item{freq}{data frequency shorthand (CMIP5 syntax, def: \code{"day"} for daily)}
}
\value{
a list object containing all of the parameters listed above, to be used
 as an input in thd other functions in this package.
}
\description{
\code{set.defaults} creates a list object that stores the default
input parameters for the package \code{quantproj}, to be
used as an input for most other functions. It also creates the
directories set as defaults (and their subdirectories used in
various bits of code in this package) if they do not yet exist.
}
\section{Directory structure}{

Generally, this package uses three main directories (though
\code{aux.dir} may be the same as either of the data directories,
the user is discouraged from setting the two data directories equal
or using pre-existing directories, for file dependency reasons):
\describe{
	\item{\code{aux_dir}}{a directory containing two sub-directories,
		\code{bases} for saved spline basis functions (the
		predictors in the quantile regression), and \code{run_logs}
		for logging progress on the bigger scripts}
	\item{\code{base.data.dir}}{the directory in which to place the raw data to
	be projected. Contains a sub-directory, \code{output}, for the resultant
	projections.}
	\item{\code{mod.data.dir}}{the directory in which to place the model data
	whose distributional changes are used to project the data in
	\code{base.data.dir}. Contains a sub-directory, \code{params}, in which the
	quantile fit parameters are stored.}
}
}

\section{Note on file structure}{

 This package assumes all files follow CMIP5 standard file naming
 conventions; in other words, it assumes that variable, data frequency, model
 name, and file year range are all encoded in the filenames themselves (see
 the
 \href{https://cmip.llnl.gov/cmip5/docs/cmip5_data_reference_syntax.pdf}{CMIP5
 syntax guide} for more information). The filename format is:
 \code{[variable]_[frequency]_[model]_[experiment]_[run/ensemble
 member]_[timeframe](_[suffix])}
}

