% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qchange_calcs_func.R
\name{get.quantile.changes}
\alias{get.quantile.changes}
\title{Get changes in estimated quantiles}
\usage{
get.quantile.changes(defaults, export.nc = T, base.years, future.years,
  load.fn = character())
}
\arguments{
\item{defaults}{the output of \code{set.defaults}}

\item{export.nc}{whether to directly save the output in a NetCDF file}

\item{base.years}{the base years to average over (i.e. \code{seq(1979,2010)})}

\item{future.years}{the future years to average over (i.e. \code{seq(2058,2099)})}

\item{load.fn}{the file containing the quantile fit coefficients. If left empty,
it will look for the output file of \code{\link{combine.locs}}, which concantenates the output of with data from
\code{\link{estimate.quantiles}}.}
}
\description{
\code{get.quantile.changes} calculates the change in the mean quantile by day
of year between a future period (mean across \code{future.years}) and a base
period (mean across \code{base.years}) by day of year for quantile estimates
generated by \code{\link{estimate.quantiles}} and merged by
\code{\link{combine.params}}). In words, for every estimated quantile and every
day of year, this code gives the difference between the average day of year
(i.e. Jan 1st) of that quantile in the \code{future.years} period vs. the
same in the \code{base.years} period.
}
\details{
Duplicate locations are removed from the final output.
}
\section{NetCDF output structure}{

 If \code{export.nc=TRUE}, the quantile changes are saved to a NetCDF file.
 The main variable, "\code{d[defaults$filevar]}", is a \code{[loc x quantile
 x day of year]} array. \code{lat}, \code{lon}, and \code{q}/quantile are also
 saved. The day of year is just an index \code{1:365}. A few extra global
 attributes are included, and an additional "comment" attribute can be added
 to the NetCDF file using the \code{comment} parameter.

 The NetCDF file is saved in \code{defaults$mod.data.dir} with the name:
 d[filevar]_day_[mod.name]_qchange_[base.year[1]-base.year[end]]_[future.year[1]-future.year[end]].nc
}

